requirements.txt
pandas
streamlit
matplotlib
seaborn
mplsoccer
statsbombpy
fastapi
transformers
torch
uvicorn
langchain_community
langchain==0.3.7
langchain-huggingface
huggingface_hub
sentencepiece
langchain_google_genai
python-dotenv
httpie